update SMt template uoload and get wyt tracking number

+++ sync/ibay_sync/upload_smt_templates.py
@@ -34,23 +34,38 @@ class Upload(BaseService):
                 data = {'mubanxls': ('report.xls', files, 'application/vnd.ms-excel')}
                 res = self.session.post(self.upload_url, files=data)
                 soup = BeautifulSoup(res.content, features='html.parser')
-                # print(soup)
-
+                table = soup.findAll(class_='ebay_table')
                 html = soup.find(text=re.compile("导入成功"))
             if html:
                     mubanId = re.findall(r'\d+', html)
                     # print(mubanId)
                     if mubanId :
                         self.remark_data(path, mubanId[0])
+            else:
+                if len(table)>1:
+                    log = table[1].findAll('td')[0].getText()
+                    self.update_log(path, log)
+
         except Exception as why:
             self.logger.error('{} is failed to upload cause of {}'.format(path, why))
-            os.remove(path)  # 删除excel文件
+            # os.remove(path)  # 删除excel文件
 
+
+    def update_log(self, path, content):
+        try:
+            if path.find('SMT1') != -1:
+                list = path.split('.')
+                sql = "update proCenter.oa_smtImportToIbayLog set content=%s where ibaySuffix=%s and sku=%s;"
+                params = (content, list[6], list[5])
+
+                self.warehouse_cur.execute(sql, params)
+                self.warehouse_con.commit()
+
+                os.remove(path)  # 删除excel文件
+        except Exception as why:
+            self.logger.error('Failed to update log content cause of {}'.format(why))
 
     def remark_data(self, path, muban_id):
-        sql = ''
-        params = ()
-        list = []
         try:
             now = str(datetime.datetime.now().strftime('%Y-%m-%d %H-%M-%S'))
             if path.find('SMT1') != -1:
@@ -58,10 +73,10 @@ class Upload(BaseService):
                 sql = "update proCenter.oa_smtImportToIbayLog set mubanId=%s,completeDate1=%s,status1=%s where ibaySuffix=%s and sku=%s;"
                 params = (muban_id, now, 1, list[6], list[5])
 
-            self.warehouse_cur.execute(sql, params)
-            self.warehouse_con.commit()
+                self.warehouse_cur.execute(sql, params)
+                self.warehouse_con.commit()
 
-            os.remove(path)  # 删除excel文件
+                os.remove(path)  # 删除excel文件
         except Exception as why:
             self.logger.error('Failed to remark data cause of {}'.format(why))
 
@@ -73,7 +88,7 @@ class Upload(BaseService):
             for input_file in os.listdir(self.path):
                 path = self.path + input_file
                 self.upload(path)  # 上传表格
-                os.remove(path)
+                # os.remove(path)
 
         except Exception as why:
             self.logger.error('Failed to upload goods templates cause of {}'.format(why))
